created ThemeToggleComponent.vue
