+++ frontend/src/App.vue
@@ -33,14 +33,18 @@ main {
   min-height: calc(100vh - 10rem);
   background:
   /* Top */
-  radial-gradient(farthest-corner at 500px 500px, var(--blurredColour1) 0.1%, rgba(255, 255, 255, 0) 26%),
+  radial-gradient(
+    farthest-corner at 500px 500px,
+    var(--blurredColour1) 0.1%,
+    rgba(255, 255, 255, 0) 26%
+  ),
   radial-gradient(farthest-corner at 500px 500px, var(--blurredColour1) 0.1%, rgba(255, 255, 255, 0) 26%),
   radial-gradient(farthest-corner at 7% 7%, var(--blurredColour2) 0.1%, rgba(255, 255, 255, 0) 8%),
   radial-gradient(farthest-corner at 17% 40px,  var(--blurredColour3) 1%, rgba(255, 255, 255, 0) 16%),
   radial-gradient(farthest-corner at 17% 40px,  var(--blurredColour3) 1%, rgba(255, 255, 255, 0) 16%),
   radial-gradient(farthest-corner at 36% 7%, var(--blurredColour4) 0.1%, rgba(255, 255, 255, 0) 18%),
   radial-gradient(farthest-corner at 36% 7%, var(--blurredColour4) 0.1%, rgba(255, 255, 255, 0) 18%),
-  
+
   /* Bottom */
   radial-gradient(farthest-corner at 95% 42%, var(--blurredColour5) 0.1%, rgba(255, 255, 255, 0) 26%),
   radial-gradient(farthest-corner at 0% 58%, var(--blurredColour5) 0.1%, rgba(255, 255, 255, 0) 22%),
@@ -51,7 +55,7 @@ main {
   background-color: var(--background-color-primary);
   backdrop-filter: blur(5px);
   color: var(--text-primary-color) !important;
-}   
+}
 
 p {
   color: var(--text-primary-color);
